+++ backend/api/measure/app.go
@@ -5152,6 +5152,14 @@ func GetSession(c *gin.Context) {
 
 	threads.Sort()
 
+	sessionTraces, err := span.FetchTracesForSessionId(ctx, appId, sessionId)
+	if err != nil {
+		msg := `failed to fetch trace data for timeline`
+		fmt.Println(msg, err)
+		c.JSON(http.StatusInternalServerError, gin.H{"error": msg})
+		return
+	}
+
 	response := gin.H{
 		"session_id":            sessionId,
 		"attribute":             session.Attribute,
@@ -5161,6 +5169,7 @@ func GetSession(c *gin.Context) {
 		"memory_usage":          memoryUsages,
 		"memory_usage_absolute": memoryUsageAbsolutes,
 		"threads":               threads,
+		"traces":                sessionTraces,
 	}
 
 	c.JSON(http.StatusOK, response)